requirements.txt
numpy>=1.20.0
pandas>=1.3.0
spacy>=3.0.0
nltk>=3.6.0
transformers>=4.18.0
pyyaml>=6.0
tqdm>=4.62.0
tensorflow-2.19.0
wheel>=0.45.1
scikit-learn>=1.6.1
sentence_transformers>=4.1.0
spicy>=1.15.2
nltk>=3.6.0
pyyaml>=6.0
tqdm>=4.62.0
matplotlib>=3.5.0
tf-keras>=2.19.0
pypdf>=5.4.0
pdfplumber>=0.11.6
pypdf2>=3.0.1
tabulate>=0.9.0
langchain_openai>=0.2.5
faiss-cpu>=1.11.0